# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 37AC 识别服务 API
  version: 0.0.1
  description: |
    37AC 角色识别服务的 HTTP 接口定义（Flask 应用 AC_web，源码在 src/server）。
    由源码 + 实测抓包整理，可直接导入 Apifox（导入数据 -> OpenAPI/Swagger）。

    约定与注意事项：
    - 所有 JSON 响应都带 success 与 code（机器可读错误码，2026-09-13 起）。历史结构一律保留：
      多数接口是 {success,message,data}；dashboard/nodes/upload/tasks 是 {type,timestamp,data...}；
      dashboard 异常分支是 {error}（另外补一份 message）。
    - code 取值 = 状态码默认码（OK/CREATED/ACCEPTED/INVALID_PARAMS/UNAUTHORIZED/FORBIDDEN/NOT_FOUND/
      METHOD_NOT_ALLOWED/PAYLOAD_TOO_LARGE/RATE_LIMITED/INTERNAL_ERROR/NOT_IMPLEMENTED/SERVICE_UNAVAILABLE）
      或业务码（AUTH_TOKEN_MISSING/AUTH_TOKEN_INVALID/AUTH_TOKEN_TYPE_INVALID/AUTH_INVALID_CREDENTIALS/
      AUTH_REFRESH_TOKEN_INVALID/PERMISSION_DENIED/API_KEY_MISSING/API_KEY_INVALID/DB_UNAVAILABLE/MAINTENANCE_MODE）。
      完整码表见 docs/api-inventory.md 第 1.1 节。
    - 鉴权分两类：Bearer JWT（用户态接口）与 X-API-Key（识别任务接口）。
    - POST /upload 受滑动窗口限流（默认 5 次/秒/IP），超限 429。
    - 维护模式开启时（settings.maintenance_mode=1），/auth、/admin、/static 之外一律 503。
    - 已知问题详见 docs/api-inventory.md 第 3 节（例如 /tasks/<id> 未校验任务归属）。
servers:
  - url: http://127.0.0.1:13138
    description: 本地/默认（Flask 直连；nginx 8000 只服务 PHP 站点，未反代 API）
...
security: []
paths:
  /auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, email]
              properties:
                username: {type: string, example: newuser}
                password: {type: string, format: password, example: "Passw0rd!"}
                email: {type: string, format: email, example: user@example.com}
      responses:
        "201":
          description: 注册成功
          content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}, example: {success: true, message: 注册成功, data: {id: 3, username: newuser}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
  /auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string, example: admin}
                password: {type: string, format: password, example: "******"}
      responses:
        "200":
          description: 登录成功
          content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}, example: {success: true, message: 登录成功, data: {user_id: 1, username: admin, role: admin, access_token: "eyJ...", refresh_token: "eyJ...", expires_in: 7200}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /auth/refresh:
    post:
      tags: [认证]
      summary: 刷新访问令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        "200": {description: 成功, content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /auth/verify:
    get:
      tags: [认证]
      summary: 校验 JWT 令牌
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 令牌有效
          content: {application/json: {example: {success: true, message: 令牌有效, data: {user_id: 1, username: admin, role: admin}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /auth/logout:
    post:
      tags: [认证]
      summary: 注销（服务端无状态，客户端丢弃令牌即可）
      security:
        - bearerAuth: []
      responses:
        "200": {description: 成功, content: {application/json: {example: {success: true, message: 注销成功}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /auth/forgot-password:
    post:
      tags: [认证]
      summary: 发送密码重置邮件
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [email], properties: {email: {type: string, format: email}}}
      responses:
        "200": {description: 已发送, content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
  /auth/reset-password:
    post:
      tags: [认证]
      summary: 用重置令牌设置新密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password, confirm_password]
              properties:
                token: {type: string}
                password: {type: string, format: password}
                confirm_password: {type: string, format: password}
      responses:
        "200": {description: 成功, content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
  /auth/verify-reset-token:
    post:
      tags: [认证]
      summary: 校验密码重置令牌是否有效
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [token], properties: {token: {type: string}}}
      responses:
        "200": {description: 有效, content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: POST /auth/login 返回的 data.access_token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: POST /api-keys 创建，形如 37ac_xxx（仅创建时返回一次）
  responses:
    BadRequest:
      description: 参数错误
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}, example: {success: false, message: 参数错误}}}
    Unauthorized:
      description: 未认证或令牌无效
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}, example: {success: false, message: 未提供认证令牌}}}
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success: {type: boolean, example: true}
        code: {type: string, example: OK, description: 机器可读错误码，见 api-inventory.md 1.1}
        message: {type: string, example: 操作成功}
        data: {type: object, nullable: true}
    ErrorEnvelope:
      type: object
      properties:
        success: {type: boolean, example: false}
        code: {type: string, example: INVALID_PARAMS, description: 机器可读错误码，见 api-inventory.md 1.1}
        message: {type: string, example: 参数错误}
...